# app/src/main/resources/static/openapi3.yml
openapi: 3.0.1
info:
  title: depromeet
  description: description
  version: '1.0.1'
servers:
  - url: https://jjaksim.com
    description: prod
  - url: http://localhost:8080
    description: local
  - url: https://1f70f9e5-cebb-4d8e-8b33-f590a74edacf.mock.pstmn.io/
    description: mock
...
paths:
  /api/v1/habits:
    post:
      tags:
        - Habit
      summary: 습관 생성 시 사용되는 API
      description: saveHabitAPI
      operationId: saveHabit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveHabitRequest'
        required: true
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Habit'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "invalid.id"
                  message: "id는 숫자여야만 합니다."
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
    get:
      tags:
        - Habit
      summary: 모든 습관 목록을 가져오는 API
      description: 모든 habit list를 반환한다
      operationId: getHabits
      parameters:
        - name: status
          in: query
          required: true
          schema:
            type: string
            enum: [ "ACTIVE", "ARCHIVED" ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HabitOverview'
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
  /api/v1/habits/{habitId}:
    get:
      tags:
        - Habit
      summary: 습관 상세정보 API (나의 기록 - 상세기록에서 특정 습관 정보 확인 시 사용, 습관 삭제 시 어떤 모달을 띄울지 판별하기 위해서도 사용합니다.)
      description: todayHabitAchievementId 가 null인 경우 오늘 수행한 기록이 없는 경우. totalAchievementCount 를 통해 그동안 수행한 기록이 있는지 판별하고, mate가 null인지 아닌지로 짝꿍 여부를 판별할 수 있습니다.
      operationId: getSingleHabit
      parameters:
        - name: habitId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Habit'
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
    put:
      tags:
        - Habit
      summary: 습관 수정 API
      operationId: updateHabit
      parameters:
        - name: habitId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditHabitRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Habit'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "invalid.id"
                  message: "id는 숫자여야만 합니다."
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
    delete:
      tags:
        - Habit
      summary: 습관 삭제 API
      operationId: deleteHabit
      parameters:
        - name: habitId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
  /api/v1/habits/{habitId}/mates:
    post:
      tags:
        - Mate
      summary: 짝꿍 생성 API
      description: 짝꿍 생성 API
      operationId: saveMate
      parameters:
        - name: habitId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MateRequest'
        required: true
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mate'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "invalid.id"
                  message: "id는 숫자여야만 합니다."
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
  /api/v1/habits/{habitId}/mates/{mateId}:
    get:
      tags:
        - Mate
      summary: 짝꿍 상세조회 API
      operationId: getMate
      parameters:
        - name: habitId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: mateId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mate'
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
    delete:
      tags:
        - Mate
      summary: 짝꿍 삭제 API
      operationId: deleteMate
      parameters:
        - name: habitId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: mateId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
  /api/v1/members:
    post:
      tags:
        - Member
      summary: 사용자의 소셜 회원가입/로그인을 진행한다.
      description: login/join
      operationId: saveMember
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Member'
                  - type: object
                    properties:
                      token:
                        $ref: '#/components/schemas/Token'
        '201':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Member'
                  - type: object
                    properties:
                      token:
                        $ref: '#/components/schemas/Token'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "invalid.id"
                  message: "id는 숫자여야만 합니다."
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
    get:
      tags:
        - Member
      summary: 회원 상세 정보 API
      operationId: getMember
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
    delete:
      tags:
        - Member
      summary: 회원 탈퇴 API
      operationId: deleteMember
      responses:
        '204':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
...
  /api/v1/members/name:
    patch:
      tags:
        - Member
      summary: 닉네임 수정 API
      operationId: patchMemberName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberNameUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "invalid.id"
                  message: "id는 숫자여야만 합니다."
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
  /api/v1/members/consents:
    patch:
      tags:
        - Member
      summary: 알림수신여부 수정 API
      operationId: patchMemberNotificationConsent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberNotificationConsentUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '400':
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "invalid.id"
                  message: "id는 숫자여야만 합니다."
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
...
  /api/v1/members/tokens:
    post:
      tags:
        - Member
      summary: 토큰 만료시 재발급을 위한 API
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Token'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '500':
          description: INTERNAL SERVER ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiFailureResponse'
              example:
                - code: "uncaught error"
                  message: "서버 에러"
...
security:
  - bearerAuth: [ token ]
components:
  securitySchemes:
    token:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Habit:
      required: [ habitId, memberId, title, imojiPath, dayOfWeeks, level, characterType, createAt, reward, notification ]
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        memberId:
          type: integer
          example: 1
        title:
          type: string
          example: "물을 마시자"
        imojiPath:
          type: string
          example: "imojiString"
        dayOfWeeks:
          type: array
          example: [ "MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY" ]
        reward:
          type: integer
          example: 3
        color:
          type: string
          example: pink
        status:
          type: string
          enum: [ "ACTIVE", "ARCHIVED" ]
          example: "ACTIVE"
        createAt:
          type: string
          example: "2022-11-30 21:40:18.668458"
        sequence:
          type: integer
          example: 0
        todayHabitAchievementId:
          type: integer
          example: 81
        totalAchievementCount:
          type: integer
          example: 0
        mate:
          $ref: '#/components/schemas/Mate'
        notification:
          $ref: '#/components/schemas/Notification'
...
    HabitOverview:
      required: [ habitId, memberId, title, imojiPath, dayOfWeeks, reward, color, createAt, todayHabitAchievementId, sequence, mate ]
      type: object
      properties:
        habitId:
          type: integer
          format: int64
          example: 1
        memberId:
          type: integer
          example: 1
        title:
          type: string
          example: "물을 마시자"
        imojiPath:
          type: string
          example: "imojiString"
        dayOfWeeks:
          type: array
          example: [ "MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY" ]
        reward:
          type: integer
          example: 0
        color:
          type: string
          example: pink
        status:
          type: string
          enum: [ "ACTIVE", "ARCHIVED" ]
          example: "ACTIVE"
        createAt:
          type: string
          example: "2022-11-30 21:40:18.668458"
        todayHabitAchievementId:
          type: integer
          example: 81
        sequence:
          type: integer
          example: 1
        mate:
          $ref: '#/components/schemas/Mate'
    SaveHabitRequest:
      required: [ title, imojiPath, dayOfWeeks, color ]
      type: object
      properties:
        title:
          type: string
          example: "물을 마시자"
        imojiPath:
          type: string
          example: "imojiString"
        color:
          type: string
          example: "pink"
        dayOfWeeks:
          type: array
          example: [ "MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY" ]
        notification:
          $ref: '#/components/schemas/Notification'
...
    EditHabitRequest:
      required: [ habitId, title, imojiPath, dayOfWeek, notification ]
      type: object
      properties:
        habitId:
          type: integer
          example: 1
        title:
          type: string
          example: "물을 마시자"
        imojiPath:
          type: string
          example: "imojiString"
        dayOfWeek:
          type: array
          example: [ "MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY" ]
        notification:
          $ref: '#/components/schemas/Notification'
    MemberRequest:
      required: [ certificationSubject, oAuthToken ]
      type: object
      properties:
        certificationSubject:
          type: string
          enum: [ "GOOGLE", "KAKAO", "APPLE" ]
        oAuthToken:
          type: string
          example: "oAuthToken"
    MemberNameUpdateRequest:
      type: object
      properties:
        name:
          type: string
          example: depromeet8manager
    MemberNotificationConsentUpdateRequest:
      type: object
      properties:
        notificationConsent:
          type: boolean
          example: true
...
    MateRequest:
      type: object
      properties:
        title:
          type: string
          example: "짝꿍이"
        characterType:
          enum: [ "WHIP", "CARROT" ]
          example: "WHIP"
...
    Member:
      type: object
      properties:
        id:
          type: integer
          example: "1"
        name:
          type: string
          example: "depromeet8manager"
        certificationSubject:
          type: string
          enum: [ "GOOGLE", "KAKAO", "APPLE" ]
        notificationConsent:
          type: boolean
          example: true
        resource:
          type: object
    Mate:
      type: object
      properties:
        id:
          type: integer
          example: 1
        memberId:
          type: integer
          example: 1
        habitId:
          type: integer
          example: 1
        title:
          type: string
          example: "짝꿍이"
        level:
          type: integer
          example: 1
        characterType:
          enum: [ "WHIP", "CARROT" ]
          example: "WHIP"
        levelUpAt:
          type: string
          format: date-time
          example: "2022-11-30 21:40:18.668458"
        createAt:
          type: string
          format: date-time
          example: "2022-11-30 21:40:18.668458"
...
    Notification:
      type: object
      nullable: true
      properties:
        notificationTime:
          type: string
          example: "21:40:18"
        contents:
          type: string
          example: "빨리 물 마셔!"
    ApiResponse:
      type: object
      xml:
        name: '##default'
    ApiFailureResponse:
      type: object
      properties:
        code:
          type: string
          example: bad.request
        message:
          type: string
          example: "id 컬럼은 숫자여야만 합니다."
      xml:
        name: '##default'
    Token:
      type: object
      properties:
        accessToken:
          type: string
          example: accessToken
        refreshToken:
          type: string
          example: refreshToken
...